move k8s-apiserver-proxy to separate binary
