+++ snap/snapcraft.yaml
@@ -55,7 +55,7 @@ parts:
       INSTALL="${SNAPCRAFT_PART_INSTALL}/bin"
       mkdir -p "${INSTALL}"
 
-      for binary in k8s k8sd k8s-dqlite dqlite; do
+      for binary in k8s k8sd k8s-dqlite k8s-apiserver-proxy dqlite; do
         cp "bin/static/${binary}" "${INSTALL}/${binary}"
       done
