+++ src/k8s/Makefile
@@ -25,7 +25,7 @@ go.k8s-dqlite.bench.dqlite:
 	CGO_LDFLAGS_ALLOW=-Wl,-z,now go test -tags=dqlite,libsqlite3 -v ./test/k8s-dqlite -run "^$$" -bench "Benchmark" -benchmem
 
 ## Static Builds
-static: bin/static/k8s bin/static/k8sd bin/static/k8s-dqlite bin/static/dqlite
+static: bin/static/k8s bin/static/k8sd bin/static/k8s-dqlite bin/static/k8s-apiserver-proxy bin/static/dqlite
 
 deps/static/lib/libdqlite.a: hack/compile-static-dqlite.sh
 	./hack/compile-static-dqlite.sh
@@ -43,6 +43,9 @@ bin/static/k8s: deps/static/lib/libdqlite.a $(GO_SOURCES)
 			-o bin/static/k8s \
 			-ldflags '-s -w -linkmode "external" -extldflags "-static"' \
 			cmd/main.go
+
+bin/static/k8s-apiserver-proxy: bin/static/k8s
+	cd bin/static && ln -sf k8s k8s-apiserver-proxy
 
 bin/static/k8s-dqlite: bin/static/k8s
 	cd bin/static && ln -sf k8s k8s-dqlite
@@ -67,7 +70,7 @@ bin/static/dqlite: deps/static/lib/libdqlite.a
 			github.com/canonical/go-dqlite/cmd/dqlite
 
 ## Dynamic Builds
-dynamic: bin/dynamic/lib/libdqlite.so bin/dynamic/k8s bin/dynamic/k8sd bin/dynamic/k8s-dqlite bin/dynamic/dqlite
+dynamic: bin/dynamic/lib/libdqlite.so bin/dynamic/k8s bin/dynamic/k8sd bin/dynamic/k8s-dqlite bin/dynamic/k8s-apiserver-proxy bin/dynamic/dqlite
 
 bin/dynamic/lib/libdqlite.so: deps/dynamic/lib/libdqlite.so
 	mkdir -p bin/dynamic/lib
@@ -87,6 +90,9 @@ bin/dynamic/k8s: deps/dynamic/lib/libdqlite.so
 			-o bin/dynamic/k8s \
 			-ldflags '-s -w' \
 			cmd/main.go
+
+bin/dynamic/k8s-apiserver-proxy: bin/dynamic/k8s
+	cd bin/dynamic && ln -s k8s k8s-apiserver-proxy
 
 bin/dynamic/k8s-dqlite: bin/dynamic/k8s
 	cd bin/dynamic && ln -s k8s k8s-dqlite
