+++ src/k8s/cmd/k8s-apiserver-proxy/main.go
@@ -1,0 +1,9 @@
+package k8s_apiserver_proxy
+
+import "os"
+
+func Main() {
+	if rootCmd.Execute() != nil {
+		os.Exit(1)
+	}
+}
