+++ src/k8s/cmd/main.go
@@ -6,14 +6,16 @@ import (
 	"path/filepath"
 
 	"github.com/canonical/k8s/cmd/k8s"
+	k8s_apiserver_proxy "github.com/canonical/k8s/cmd/k8s-apiserver-proxy"
 	k8s_dqlite "github.com/canonical/k8s/cmd/k8s-dqlite"
 	"github.com/canonical/k8s/cmd/k8sd"
 	"github.com/docker/docker/pkg/reexec"
 )
 
 func init() {
-	reexec.Register("k8s", k8s.Main)
+	reexec.Register("k8s-apiserver-proxy", k8s_apiserver_proxy.Main)
 	reexec.Register("k8s-dqlite", k8s_dqlite.Main)
+	reexec.Register("k8s", k8s.Main)
 	reexec.Register("k8sd", k8sd.Main)
 }
